requirements.txt
fastapi
uvicorn
pytest
pytest-cov
pylint
python-dotenv
mysql-connector-python
sqlalchemy
sqlalchemy-utils
pika
aio-pika